refactor: PathFinderFacade를 추가한다
